@@ -1,16 +1,17 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RootNamespace>BlazorHero.CleanArchitecture.Infrastructure.Shared</RootNamespace>
     <AssemblyName>BlazorHero.CleanArchitecture.Infrastructure.Shared</AssemblyName>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Application\Application.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="MailKit" Version="2.13.0" />
-    <PackageReference Include="MimeKit" Version="2.13.0" />
+    <PackageReference Include="MailKit" Version="3.1.0" />
+    <PackageReference Include="MimeKit" Version="3.1.0" />
   </ItemGroup>
 </Project>
