@@ -19,11 +19,14 @@ using BlazorHero.CleanArchitecture.Server.Localization;
 using BlazorHero.CleanArchitecture.Server.Managers.Preferences;
 using BlazorHero.CleanArchitecture.Server.Services;
 using BlazorHero.CleanArchitecture.Server.Settings;
+using BlazorHero.CleanArchitecture.Shared.Constants.Application;
 using BlazorHero.CleanArchitecture.Shared.Constants.Localization;
 using BlazorHero.CleanArchitecture.Shared.Constants.Permission;
 using BlazorHero.CleanArchitecture.Shared.Wrapper;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
@@ -33,6 +36,7 @@ using Microsoft.Extensions.Localization;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using Newtonsoft.Json;
+using Serilog;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -57,6 +61,42 @@ namespace BlazorHero.CleanArchitecture.Server.Extensions
             return localizer;
         }
 
+        internal static IServiceCollection AddForwarding(this IServiceCollection services, IConfiguration configuration)
+        {
+            var applicationSettingsConfiguration = configuration.GetSection(nameof(AppConfiguration));
+            var config = applicationSettingsConfiguration.Get<AppConfiguration>(); 
+            if (config.BehindSSLProxy)
+            {
+                services.Configure<ForwardedHeadersOptions>(options =>
+                {
+                    options.ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto;
+                    if (!string.IsNullOrWhiteSpace(config.ProxyIP))
+                    {
+                        var ipCheck = config.ProxyIP;
+                        if (IPAddress.TryParse(ipCheck, out var proxyIP))
+                            options.KnownProxies.Add(proxyIP);
+                        else
+                            Log.Logger.Warning("Invalid Proxy IP of {IpCheck}, Not Loaded", ipCheck);
+                    }
+                });
+
+                services.AddCors(options =>
+                {
+                    options.AddDefaultPolicy(
+                        builder =>
+                        {
+                            builder
+                                .AllowCredentials()
+                                .AllowAnyHeader()
+                                .AllowAnyMethod()
+                                .WithOrigins(config.ApplicationUrl.TrimEnd('/'));
+                        });
+                });
+            }
+            
+            return services;
+        }
+
         private static async Task SetCultureFromServerPreferenceAsync(IServiceProvider serviceProvider)
         {
             var storageService = serviceProvider.GetService<ServerPreferenceManager>();
@@ -64,11 +104,10 @@ namespace BlazorHero.CleanArchitecture.Server.Extensions
             {
                 // TODO - should implement ServerStorageProvider to work correctly!
                 CultureInfo culture;
-                var preference = await storageService.GetPreference() as ServerPreference;
-                if (preference != null)
-                    culture = new CultureInfo(preference.LanguageCode);
+                if (await storageService.GetPreference() is ServerPreference preference)
+                    culture = new(preference.LanguageCode);
                 else
-                    culture = new CultureInfo(LocalizationConstants.SupportedLanguages.FirstOrDefault()?.Code ?? "en-US");
+                    culture = new(LocalizationConstants.SupportedLanguages.FirstOrDefault()?.Code ?? "en-US");
                 CultureInfo.DefaultThreadCurrentCulture = culture;
                 CultureInfo.DefaultThreadCurrentUICulture = culture;
                 CultureInfo.CurrentCulture = culture;
@@ -228,7 +267,7 @@ namespace BlazorHero.CleanArchitecture.Server.Extensions
         internal static IServiceCollection AddJwtAuthentication(
             this IServiceCollection services, AppConfiguration config)
         {
-            var key = Encoding.ASCII.GetBytes(config.Secret);
+            var key = Encoding.UTF8.GetBytes(config.Secret);
             services
                 .AddAuthentication(authentication =>
                 {
@@ -253,6 +292,20 @@ namespace BlazorHero.CleanArchitecture.Server.Extensions
 
                     bearer.Events = new JwtBearerEvents
                     {
+                        OnMessageReceived = context =>
+                        {
+                            var accessToken = context.Request.Query["access_token"];
+
+                            // If the request is for our hub...
+                            var path = context.HttpContext.Request.Path;
+                            if (!string.IsNullOrEmpty(accessToken) &&
+                                (path.StartsWithSegments(ApplicationConstants.SignalR.HubUrl)))
+                            {
+                                // Read the token out of the query string
+                                context.Token = accessToken;
+                            }
+                            return Task.CompletedTask;
+                        },
                         OnAuthenticationFailed = c =>
                         {
                             if (c.Exception is SecurityTokenExpiredException)
