@@ -27,18 +27,35 @@ namespace BlazorHero.CleanArchitecture.Application.Features.ExtendedAttributes.C
             where TExtendedAttribute : AuditableEntityExtendedAttribute<TId, TEntityId, TEntity>, IEntity<TId>
             where TId : IEquatable<TId>
     {
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public TId Id { get; set; }
+#pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public TEntityId EntityId { get; set; }
+#pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
         public EntityExtendedAttributeType Type { get; set; }
+
         [Required(AllowEmptyStrings = false)]
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public string Key { get; set; }
+#pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
         public string? Text { get; set; }
+
         public decimal? Decimal { get; set; }
+
         public DateTime? DateTime { get; set; }
+
         public string? Json { get; set; }
+
         public string? ExternalId { get; set; }
+
         public string? Group { get; set; }
+
         public string? Description { get; set; }
+
         public bool IsActive { get; set; }
     }
 
@@ -64,7 +81,7 @@ namespace BlazorHero.CleanArchitecture.Application.Features.ExtendedAttributes.C
 
         public async Task<Result<TId>> Handle(AddEditExtendedAttributeCommand<TId, TEntityId, TEntity, TExtendedAttribute> command, CancellationToken cancellationToken)
         {
-            if (await _unitOfWork.Repository<TExtendedAttribute>().Entities.Where(x => !x.Id.Equals(command.Id) && x.EntityId.Equals(command.EntityId))
+            if (await _unitOfWork.Repository<TExtendedAttribute>().Entities.Where(x => !x.Id.Equals(command.Id) && x.EntityId!.Equals(command.EntityId))
                 .AnyAsync(p => p.Key == command.Key, cancellationToken))
             {
                 return await Result<TId>.FailAsync(_localizer["Extended Attribute with this Key already exists."]);
