@@ -56,7 +56,7 @@ namespace BlazorHero.CleanArchitecture.Client.Pages.Catalog
             await GetBrandsAsync();
             _loaded = true;
 
-            HubConnection = HubConnection.TryInitialize(_navigationManager);
+            HubConnection = HubConnection.TryInitialize(_navigationManager, _localStorage);
             if (HubConnection.State == HubConnectionState.Disconnected)
             {
                 await HubConnection.StartAsync();
