@@ -20,7 +20,7 @@ namespace BlazorHero.CleanArchitecture.Client.Shared.Components
 
         private FluentValidationValidator _fluentValidationValidator;
         private bool Validated => _fluentValidationValidator.Validate(options => { options.IncludeAllRuleSets(); });
-
+        private bool _uploading = false;
         public void Cancel()
         {
             MudDialog.Cancel();
@@ -30,6 +30,7 @@ namespace BlazorHero.CleanArchitecture.Client.Shared.Components
         {
             if (OnSaved != null)
             {
+                _uploading = true;
                 var result = await OnSaved.Invoke(UploadRequest);
                 if (result.Succeeded)
                 {
@@ -43,6 +44,7 @@ namespace BlazorHero.CleanArchitecture.Client.Shared.Components
                         _snackBar.Add(message, Severity.Error);
                     }
                 }
+                _uploading = false;
             }
             else
             {
