@@ -1,8 +1,10 @@
 ﻿@using BlazorHero.CleanArchitecture.Shared.Constants.Localization
+@inject Microsoft.Extensions.Localization.IStringLocalizer<LanguageSelector> _localizer
+
 <MudMenu Icon="@Icons.Material.Outlined.Translate" Color="Color.Inherit" Direction="Direction.Bottom" OffsetY="true" Dense="true">
     @foreach (var language in LocalizationConstants.SupportedLanguages)
     {
-        <MudMenuItem OnClick="(()=> ChangeLanguageAsync(language.Code))">@language.DisplayName</MudMenuItem>
+        <MudMenuItem OnClick="(()=> ChangeLanguageAsync(language.Code))">@_localizer[language.DisplayName]</MudMenuItem>
     }
 </MudMenu>
 
