@@ -49,11 +49,9 @@ namespace BlazorHero.CleanArchitecture.Client.Shared
 
         protected override async Task OnInitializedAsync()
         {
-            await LoadDataAsync();
-
             _rightToLeft = await _clientPreferenceManager.IsRTL();
             _interceptor.RegisterEvent();
-            hubConnection = hubConnection.TryInitialize(_navigationManager);
+            hubConnection = hubConnection.TryInitialize(_navigationManager, _localStorage);
             await hubConnection.StartAsync();
             hubConnection.On<string, string, string>(ApplicationConstants.SignalR.ReceiveChatNotification, (message, receiverUserId, senderUserId) =>
             {
@@ -116,10 +114,23 @@ namespace BlazorHero.CleanArchitecture.Client.Shared
                     }
                 }
             });
+            hubConnection.On<string>(ApplicationConstants.SignalR.PingRequest, async (userName) =>
+            {
+                await hubConnection.SendAsync(ApplicationConstants.SignalR.PingResponse, CurrentUserId, userName);
+
+            });
 
             await hubConnection.SendAsync(ApplicationConstants.SignalR.OnConnect, CurrentUserId);
 
             _snackBar.Add(string.Format(_localizer["Welcome {0}"], FirstName), Severity.Success);
+        }
+
+        protected override async Task OnAfterRenderAsync(bool firstRender)
+        {
+            if (firstRender)
+            {
+                await LoadDataAsync();
+            }
         }
 
         private async Task LoadDataAsync()
@@ -129,25 +140,37 @@ namespace BlazorHero.CleanArchitecture.Client.Shared
             if (user == null) return;
             if (user.Identity?.IsAuthenticated == true)
             {
-                CurrentUserId = user.GetUserId();
-                FirstName = user.GetFirstName();
-                if (FirstName.Length > 0)
+                if (string.IsNullOrEmpty(CurrentUserId))
                 {
-                    FirstLetterOfName = FirstName[0];
-                }
-                SecondName = user.GetLastName();
-                Email = user.GetEmail();
-                var imageResponse = await _accountManager.GetProfilePictureAsync(CurrentUserId);
-                if (imageResponse.Succeeded)
-                {
-                    ImageDataUrl = imageResponse.Data;
-                }
+                    CurrentUserId = user.GetUserId();
+                    FirstName = user.GetFirstName();
+                    if (FirstName.Length > 0)
+                    {
+                        FirstLetterOfName = FirstName[0];
+                    }
 
-                var currentUserResult = await _userManager.GetAsync(CurrentUserId);
-                if (!currentUserResult.Succeeded || currentUserResult.Data == null)
-                {
-                    _snackBar.Add(_localizer["You are logged out because the user with your Token has been deleted."], Severity.Error);
-                    await _authenticationManager.Logout();
+                    SecondName = user.GetLastName();
+                    Email = user.GetEmail();
+                    var imageResponse = await _accountManager.GetProfilePictureAsync(CurrentUserId);
+                    if (imageResponse.Succeeded)
+                    {
+                        ImageDataUrl = imageResponse.Data;
+                    }
+
+                    var currentUserResult = await _userManager.GetAsync(CurrentUserId);
+                    if (!currentUserResult.Succeeded || currentUserResult.Data == null)
+                    {
+                        _snackBar.Add(
+                            _localizer["You are logged out because the user with your Token has been deleted."],
+                            Severity.Error);
+                        CurrentUserId = string.Empty;
+                        ImageDataUrl = string.Empty;
+                        FirstName = string.Empty;
+                        SecondName = string.Empty;
+                        Email = string.Empty;
+                        FirstLetterOfName = char.MinValue;
+                        await _authenticationManager.Logout();
+                    }
                 }
             }
         }
