@@ -1,18 +1,17 @@
 ﻿@inherits LayoutComponentBase
 @inject Microsoft.Extensions.Localization.IStringLocalizer<MainLayout> _localizer
 
-    <audio id="notification" src="/media/notification.mp3" />
+<MudRTLProvider RightToLeft="@_rightToLeft">
+    <audio id="notification" src="/media/notification.mp3"/>
     <MudThemeProvider Theme="_currentTheme" />
     <MudDialogProvider />
     <MudSnackbarProvider />
     <AuthorizeView>
         <NotAuthorized>
             <MudAppBar Elevation="4" Color="Color.Primary">
-                <MudAppBarSpacer />
+                <MudSpacer />
                 <LanguageSelector />
-                <MudTooltip Text="@_localizer["Toggle Dark Mode"]">
                     <MudIconButton Icon="@Icons.Material.Filled.Brightness4" Color="Color.Inherit" OnClick="@((e) => DarkMode())" />
-                </MudTooltip>
             </MudAppBar>
 
             <MudContainer MaxWidth="MaxWidth.Small" Class="d-flex align-center" Style="height: 100vh;">
@@ -22,10 +21,11 @@
             </MudContainer>
         </NotAuthorized>
         <Authorized>
-            <MudLayout RightToLeft="@_rightToLeft">
+            <MudLayout>
                 <MainBody OnDarkModeToggle="@DarkMode" OnRightToLeftToggle="@RightToLeftToggle">
                     @Body
                 </MainBody>
             </MudLayout>
         </Authorized>
     </AuthorizeView>
+</MudRTLProvider>
