@@ -31,8 +31,7 @@ namespace BlazorHero.CleanArchitecture.Server
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddCors();
-            services.AddSignalR();
+            services.AddForwarding(_configuration);
             services.AddLocalization(options =>
             {
                 options.ResourcesPath = "Resources";
@@ -45,6 +44,7 @@ namespace BlazorHero.CleanArchitecture.Server
             services.AddServerLocalization();
             services.AddIdentity();
             services.AddJwtAuthentication(services.GetApplicationSettings(_configuration));
+            services.AddSignalR();
             services.AddApplicationLayer();
             services.AddApplicationServices();
             services.AddRepositories();
@@ -69,7 +69,7 @@ namespace BlazorHero.CleanArchitecture.Server
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IStringLocalizer<Startup> localizer)
         {
-            app.UseCors();
+            app.UseForwarding(_configuration);
             app.UseExceptionHandling(env);
             app.UseHttpsRedirection();
             app.UseMiddleware<ErrorHandlerMiddleware>();
